requirements.txt
djangorestframework==3.14.0
Django==4.0.4
django-cors-headers==3.11.0
mysql-connector-python==8.0.28
mysqlclient==2.1.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.4
django-dotenv==1.4.2
redis==4.3.3
pyjwt==2.4.0
pandas
more_itertools
boto3
openpyxl
dynamodb_json
psutil
scikit-learn 
statsmodels
polars
pyarrow